% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_abj_book}
\alias{setup_abj_book}
\title{Setup ABJ book}
\usage{
setup_abj_book(name, path = "./")
}
\arguments{
\item{name}{nome do livro.}

\item{path}{pasta onde os arquivos serao copiados.}
}
\description{
Setupar ABJ book
}
